Added RSpec shared contexts to library
Necessary for downstream projects to include, as desired, without
having to set up these contexts individually or duplicate code.

Includes a Git repository and commit shared context for testing
purposes. The repository was extracted from the [Git
Lint](https://www.alchemists.io/projects/git-lint) while the commit
matches the initial commit created in the repository context. The two
contexts are not combined because there can be situations where only
the commit is required but not the repository which speeds up RSpec
performance.

These contexts will need to be explicitly required in downstream
projects as they are optional by default.